anykernel.sh: Return the first selection from find
Some Devices like deb have 2 fstab files. (In which I don't know why...)
When using the find command, we will return 2 results and the flash may not
successfully happen like wanted initially.

In this case, return only the first occurence of the find command. At the same
time, remove the device variable as it isn't really needed for the kernel
installation.