Balance CPU utilization
With the coming of Cpu-Boost, all 4 cores of the Snapdragon S4 Pro were
activated. This SOC isn't configured like it's newer counterparts and can't
be used as such. (Big.Little architecture)

Disable by default the 3rd and the 4th core and compensate the possible
performance loss by adjusting the 918Mhz to 1026Mhz. Apply the same thing
to the performances profiles as well.